+++ packages/temporal/workflows/mtv/mtv_actions.go
@@ -1,4 +1,4 @@
-package workflows
+package mtv
 
 import (
 	"time"
