+++ packages/temporal/workflows/mtv/mtv_events.go
@@ -1,4 +1,4 @@
-package workflows
+package mtv
 
 import (
 	"github.com/AdonisEnProvence/MusicRoom/shared"
