+++ packages/temporal/workflows/mtv/mtv_test.go
@@ -1,4 +1,4 @@
-package workflows
+package mtv
 
 import (
 	"errors"
